randomise sent data & change temperature field type to double

+++ pscf-demo/src/main/java/pl/polsl/pscfdemo/dto/InputBrokerDto.java
@@ -12,7 +12,7 @@ import java.time.Instant;
 public class InputBrokerDto {
 
 	@JsonProperty("pH")
-	Integer phValue;
+	Double phValue;
 
 	@JsonProperty("temp")
 	Double temperature;
